// backend/src/infrastructure/database/kysely/db.types.ts
export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export type Json = JsonValue;

export type JsonArray = JsonValue[];

export type JsonObject = {
  [x: string]: JsonValue | undefined;
};

export type JsonPrimitive = boolean | number | string | null;

export type JsonValue = JsonArray | JsonObject | JsonPrimitive;

export type Numeric = ColumnType<string, number | string, number | string>;

export type Timestamp = ColumnType<Date, Date | string, Date | string>;
// ...
export interface SalesOrderLines {
  delivered_quantity: Generated<Numeric>;
  discount_percent: Generated<Numeric>;
  gst_id: string | null;
  id: Generated<string>;
  line_total: Generated<Numeric>;
  product_id: string;
  product_unit_id: string;
  quantity: Numeric;
  rate: Numeric;
  so_id: string;
}

export interface SalesOrders {
  approved_at: Timestamp | null;
  approved_by: string | null;
  created_at: Generated<Timestamp>;
  created_by: string | null;
  customer_id: string;
  expected_delivery_date: Timestamp | null;
  financial_year_id: string;
  id: Generated<string>;
  quotation_id: string | null;
  remarks: string | null;
  so_date: Generated<Timestamp>;
  so_number: string;
  status: Generated<string>;
  total_amount: Generated<Numeric>;
  updated_at: Generated<Timestamp>;
  warehouse_id: string;
}